refactor: Удалить функцию-обёртку
